[#1] move files around to make more sense and be cleaner
