+++ CMakeLists.txt
@@ -28,32 +28,29 @@ FetchContent_Declare(
 )
 FetchContent_MakeAvailable(gl3w)
 
+file(GLOB INCL ${PROJECT_SOURCE_DIR}/include/math/*)
+file(GLOB IMPL ${PROJECT_SOURCE_DIR}/include/math/impl/*)
+
 enable_testing()
 
 add_executable(vec_test test/vec_test.cpp ${IMPL} ${INCL})
 
 target_include_directories(vec_test PUBLIC ${PROJECT_SOURCE_DIR}/include)
-target_include_directories(vec_test PRIVATE ${PROJECT_SOURCE_DIR}/src/math)
 
 target_link_libraries(vec_test Catch2::Catch2WithMain)
 
 add_executable(mat_test test/mat_test.cpp ${IMPL} ${INCL})
 
 target_include_directories(mat_test PUBLIC ${PROJECT_SOURCE_DIR}/include)
-target_include_directories(mat_test PRIVATE ${PROJECT_SOURCE_DIR}/src/math)
 
 target_link_libraries(mat_test Catch2::Catch2WithMain)
 
 add_test(NAME vec_test COMMAND vec_test)
 add_test(NAME mat_test COMMAND mat_test)
 
-file(GLOB INCL ${PROJECT_SOURCE_DIR}/include/**)
-file(GLOB IMPL ${PROJECT_SOURCE_DIR}/src/math/*)
-
 add_executable(${PROJECT_NAME} main.cpp ${INCL} ${IMPL})
 add_dependencies(${PROJECT_NAME} vec_test mat_test)
 
 target_include_directories(${PROJECT_NAME} PUBLIC ${PROJECT_SOURCE_DIR}/include)
-target_include_directories(${PROJECT_NAME} PRIVATE ${PROJECT_SOURCE_DIR}/src/math)
 
 target_link_libraries(${PROJECT_NAME} glfw gl3w)
