+++ include/math/impl/storage.inl
@@ -1,4 +1,4 @@
-#include "math/storage.hpp"
+#include "storage.hpp"
 
 #include <cassert>
