+++ include/math/impl/vec3_type.hpp
