+++ include/math/vec.hpp
@@ -1,8 +1,8 @@
 #pragma once
 
-#include "vec2_type.hpp"
-#include "vec3_type.hpp"
-#include "vec4_type.hpp"
+#include "impl/vec2_type.hpp"
+#include "impl/vec3_type.hpp"
+#include "impl/vec4_type.hpp"
 
 using vec2 = vec<float, 2>;
 using vec3 = vec<float, 3>;
